# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    description: Error response with status, code, and description
    properties:
      code:
        example: ERR_INVALID_REQUEST
        type: string
      description:
        example: Invalid request parameters
        type: string
      status:
        example: error
        type: string
    type: object
  handlers.GetPermissionResponse:
    description: Response containing permission details
    properties:
      description:
        example: Permission setting found.
        type: string
      permission:
        $ref: '#/definitions/handlers.PermissionDetail'
      status:
        example: success
        type: string
    type: object
...
  handlers.ListPermissionsResponse:
    description: Response containing list of permissions
    properties:
      permissions:
        items:
          $ref: '#/definitions/handlers.PermissionDetail'
        type: array
      status:
        example: success
        type: string
      totalCount:
        example: 10
        type: integer
    type: object
...
  handlers.PermissionDetail:
    description: Permission details
    properties:
      createdAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      messageBox:
        example: inbox
        type: string
      recipientFee:
        example: 100
        type: integer
      sender:
        example: 03abc...
        type: string
      status:
        example: payment_required
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
  handlers.QuoteEntry:
    description: Quote for one recipient in batch
    properties:
      deliveryFee:
        example: 10
        type: integer
      messageBox:
        example: inbox
        type: string
      recipient:
        example: 03abc...
        type: string
      recipientFee:
        example: 100
        type: integer
      status:
        example: payment_required
        type: string
    type: object
  handlers.QuoteMultiResponse:
    description: Response containing quotes for multiple recipients
    properties:
      blockedRecipients:
        items:
          type: string
        type: array
      description:
        example: Message delivery quotes generated for 2 recipients.
        type: string
      quotesByRecipient:
        items:
          $ref: '#/definitions/handlers.QuoteEntry'
        type: array
      status:
        example: success
        type: string
      totals:
        $ref: '#/definitions/handlers.QuoteTotals'
    type: object
...
  handlers.QuoteTotals:
    description: Totals for batch quote
    properties:
      deliveryFees:
        example: 20
        type: integer
      recipientFees:
        example: 200
        type: integer
      totalForPayableRecipients:
        example: 220
        type: integer
    type: object
...
  handlers.SetPermissionRequest:
    description: Request to set a permission
    properties:
      messageBox:
        example: inbox
        type: string
      recipientFee:
        example: 100
        type: integer
      sender:
        example: 03abc...
        type: string
    type: object
  handlers.SetPermissionResponse:
    description: Response after setting a permission
    properties:
      description:
        example: Messages from sender to inbox now require 100 satoshis.
        type: string
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for message delivery, retrieval, acknowledgement and permissions.
    Uses BRC-31/BRC-104 mutual authentication.
  title: MessageBox Server API
  version: 1.0.0
paths:
...
  /permissions/get:
    get:
      description: Retrieves the permission setting for a specific sender or box-wide
        default for the authenticated identity.
      parameters:
      - description: Name of the message box
        in: query
        name: messageBox
        required: true
        type: string
      - description: Sender's public key (omit for box-wide setting)
        in: query
        name: sender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Get a message permission
      tags:
      - Permissions
  /permissions/list:
    get:
      description: Returns all permission settings for the authenticated identity,
        optionally filtered by message box.
      parameters:
      - description: Filter by message box name
        in: query
        name: messageBox
        type: string
      - description: Maximum number of results (1-1000, default 100)
        in: query
        name: limit
        type: integer
      - description: Number of results to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: 'Sort order: ''asc'' or ''desc'' (default ''desc'')'
        in: query
        name: createdAtOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: List message permissions
      tags:
      - Permissions
  /permissions/quote:
    get:
      description: Returns fee information for sending a message to one or more recipients.
        Single recipient returns QuoteSingleResponse, multiple recipients returns
        QuoteMultiResponse.
      parameters:
      - description: Recipient public key (can be repeated for multiple recipients)
        in: query
        name: recipient
        required: true
        type: string
      - description: Name of the message box
        in: query
        name: messageBox
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Multiple recipient quote
          schema:
            $ref: '#/definitions/handlers.QuoteMultiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Get a delivery quote
      tags:
      - Permissions
  /permissions/set:
    post:
      consumes:
      - application/json
      description: Sets fee requirements for receiving messages. Use recipientFee=0
        for free, recipientFee=-1 to block, or a positive value for required payment
        in satoshis. Omit sender for box-wide defaults.
      parameters:
      - description: Permission settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SetPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Set a message permission
      tags:
      - Permissions
...
securityDefinitions:
  BSVAuth:
    description: BRC-31/BRC-104 mutual authentication. Requires multiple x-bsv-auth-*
      headers (identity-key, nonce, signature, etc.)
    in: header
    name: x-bsv-auth-identity-key
    type: apiKey
swagger: "2.0"